🔨 Consertando erros

+++ app/models/Questao.js
@@ -5,7 +5,7 @@ class Questao extends Model{
     static init(connection){
         super.init({
             descricao:DataTypes.STRING,
-            ativo:DataTypes.TINYINT,
+            ativo:DataTypes.INTEGER,
             created_at:DataTypes.DATE,
             updated_at:DataTypes.DATE
         },
